listener: address negotiation review feedback
